src/requirements.txt
altgraph==0.17.4
certifi==2025.7.9
charset-normalizer==3.4.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
psutil==7.0.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pyjwt==2.10.1
pyperclip==1.9.0
pywin32-ctypes==0.2.3
requests==2.32.4
setuptools==80.9.0
urllib3==2.5.0
